+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@ repos:
     hooks:
     -   id: end-of-file-fixer
     -   id: trailing-whitespace
-        exclude: src/__test__/__snapshots__
+        exclude: test/__snapshots__
 -   repo: git://github.com/Lucas-C/pre-commit-hooks
     rev: v1.1.4
     hooks:
