+++ package.json
@@ -6,7 +6,7 @@
   "main": "src/js/index.jsx",
   "scripts": {
     "build": "webpack-cli --mode production",
-    "lint:js": "eslint src",
+    "lint:js": "eslint src test",
     "start": "webpack-dev-server --open --watch --mode development --output-path build",
     "pip:install": "pip install --user -r .scripts/python_requirements.txt",
     "test": "jest",
@@ -115,7 +115,7 @@
       "jest-junit"
     ],
     "setupFilesAfterEnv": [
-      "<rootDir>/src/__test__/setup.js"
+      "<rootDir>/test/setup.js"
     ],
     "snapshotSerializers": [
       "enzyme-to-json/serializer"
