+++ sonar-project.properties
@@ -4,7 +4,7 @@ sonar.projectName=tech-radar
 sonar.sourceEncoding=UTF-8
 
 sonar.sources=src
-sonar.exclusions=src/js/lib/**,src/__test__/**
-sonar.tests=src/__test__
+sonar.exclusions=src/js/lib/**,src/**/__mocks__/**
+sonar.tests=test
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
 sonar.testExecutionReportPaths=test-results/test-report.xml
