+++ test/Footer.test.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Footer from '../js/components/Footer';
+import Footer from '../src/js/components/Footer';
 
-jest.mock('../js/modules/radarService');
+jest.mock('../src/js/modules/radarService');
 
 describe('<Footer />', () => {
   test('It renders the footer', () => {
