+++ test/radarService.test.js
@@ -1,16 +1,16 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
-import { build } from '../js/builder/builder';
-import radarService from '../js/modules/radarService';
+import { build } from '../src/js/builder/builder';
+import radarService from '../src/js/modules/radarService';
 
-jest.dontMock('../js/modules/radarService');
+jest.dontMock('../src/js/modules/radarService');
 
 describe('RadarService', () => {
   beforeAll(async () => {
     await build(path.resolve(__dirname, 'radar/valid'));
 
     global.fetch = jest.fn().mockImplementationOnce(() => Promise.resolve({
-      json: () => fs.readFile(path.resolve(__dirname, '../../build/js/radar.json'), 'utf-8')
+      json: () => fs.readFile(path.resolve(__dirname, '../build/js/radar.json'), 'utf-8')
         .then(data => JSON.parse(data)),
     }));
